+++ src/commands/hardis/org/monitor/backup.ts
@@ -103,7 +103,12 @@ export default class MonitorBackup extends SfdxCommand {
         debug: this.debugMode,
       });
     } catch (e) {
-      uxLog(this, c.yellow("Crash during backup. You may exclude more items by customizing file manifest/package-skip-items.xml"));
+      const failedPackageXmlContent = await fs.readFile(packageXmlBackUpItemsFile, "utf8");
+      uxLog(this, c.yellow("BackUp package.xml that failed to be retrieved:\n" + c.grey(failedPackageXmlContent)));
+      uxLog(
+        this,
+        c.red("Crash during backup. You may exclude more metadata types by updating file manifest/package-skip-items.xml then commit and push it"),
+      );
       throw e;
     }